requirements.txt
starlette==0.27.0
setuptools==65.5.1
pydantic==1.10.13
fastapi==0.100.0
uvicorn[standard]==0.20.0
python-dotenv
langchain==0.2.6
langchain-community==0.2.6
langchain_groq
langchain-openai==0.1.8
langchain-google-genai
google-generativeai
redis>=5.0.1